student_ready_exams

+++ src/logic/exam/DescriptiveQues.java
@@ -6,7 +6,7 @@ public class DescriptiveQues extends Question{
     public DescriptiveQues(double point, String questionText) {
         super(point, questionText);
     }
-    public DescriptiveQues(int time, int numQue, double point, String questionText) {
+    public DescriptiveQues(double time, int numQue, double point, String questionText) {
         super(time, numQue, point, questionText);
     }
