+++ src/logic/exam/True_False_ques.java
@@ -7,7 +7,7 @@ public class True_False_ques extends Question{
         super(point, questionText);
     }
 
-    public True_False_ques(int time, int numQue, double point, String questionText) {
+    public True_False_ques(double time, int numQue, double point, String questionText) {
         super(time, numQue, point, questionText);
     }
